+++ modules/App/assets/js/app.js
@@ -4,6 +4,7 @@ import "../vendor/kiss/lib.js"
 import "../vendor/storage.js";
 import "../vendor/i18n.js";
 import "../vendor/dompurify.js";
+import "../vendor/mousetrap.js";
 import utils from "./app/utils.js";
 import ui from "./app/ui.js";
 import assets from "./app/assets.js";