+++ modules/Content/admin.php
@@ -31,3 +31,28 @@ $this->on('app.permissions.collect', function($permissions) {
         ]
     ];
 });
+
+$this->on('app.search', function($search, $findings) {
+
+    $models = array_values($this->module('content')->models());
+
+    foreach ($models as $model) {
+
+        if (stripos($model['name'], $search) !== false || stripos($model['label'], $search) !== false) {
+
+            if (!$this->helper('acl')->isAllowed("content/{$model['name']}/read")) {
+                continue;
+            }
+
+            $icon = $model['type'] == 'collection' ? 'content:assets/icons/collection.svg' : 'content:assets/icons/singleton.svg';
+
+            $findings[] = [
+                'title' => isset($model['label']) && $model['label'] ? "{$model['label']} ({$model['name']})" : $model['name'],
+                'route' => $this->routeUrl($model['type'] == 'collection' ? "/content/collection/items/{$model['name']}" : "/content/singleton/item/{$model['name']}"),
+                'group' => 'Content',
+                'icon' => $icon
+            ];
+        }
+    }
+
+});